Implement PyPI package distribution support
- Add pyproject.toml with setuptools configuration for PyPI publishing
- Restructure codebase to proper Python package (src/whisper_key/)
- Move all modules, assets, and config files to package directory
- Update imports to use relative imports within package
- Create backward-compatible wrapper for whisper-key.py script
- Make ten-vad dependency optional with graceful fallback
- Update PyInstaller spec file for new package structure
- Implement dual-mode asset resolution (PyInstaller + pip/pipx)
- Add console script entry point for global whisper-key command
- Successfully tested with pipx installation and global command availability
