+++ documentation/implementation-plans/2025-08-19-1130-pypi-integration-implementation.md
@@ -27,12 +27,29 @@ As a *developer/advanced user* I want **to install whisper-key via pip/pipx** so
 ## Implementation Plan
 
 ### Phase 1: Package Structure Setup
-- [ ] Create `pyproject.toml` with minimal setuptools configuration
-- [ ] Create `src/whisper_key/__init__.py` package initialization
-- [ ] Move `whisper-key.py` to `src/whisper_key/main.py`
-- [ ] Move assets and config.defaults.yaml to package directory
-- [ ] Update imports in moved main.py file
-- [ ] Address ten-vad git dependency (make optional or vendor)
+- [x] Create `pyproject.toml` with minimal setuptools configuration
+  - ✅ Created pyproject.toml with setuptools configuration and Windows platform markers
+  - ✅ Defined console script entry point as `whisper-key = "whisper_key.main:main"`
+  - ✅ Configured package data to include assets and config files
+- [x] Create `src/whisper_key/__init__.py` package initialization
+  - ✅ Created package initialization with version management (__version__ = "0.2.0")
+  - ✅ Added package metadata and description
+- [x] Move `whisper-key.py` to `src/whisper_key/main.py`
+  - ✅ Successfully moved main entry point to package structure
+  - ✅ Maintained all existing functionality and setup functions
+- [x] Move assets and config.defaults.yaml to package directory
+  - ✅ Moved assets/ directory to src/whisper_key/assets/
+  - ✅ Moved config.defaults.yaml to src/whisper_key/config.defaults.yaml
+  - ✅ Moved all Python modules from src/ to src/whisper_key/
+- [x] Update imports in moved main.py file
+  - ✅ Updated all imports from `src.module` to relative imports `.module`
+  - ✅ Fixed cross-module imports in whisper_engine.py
+  - ✅ Verified no remaining src imports exist in codebase
+- [x] Address ten-vad git dependency (make optional or vendor)
+  - ✅ Made ten-vad import optional with try/except ImportError handling
+  - ✅ Added HAS_TEN_VAD flag for conditional functionality
+  - ✅ Updated VAD initialization to gracefully handle missing ten-vad
+  - ✅ Added warning log when VAD is enabled but ten-vad unavailable
 
 ### Phase 2: Asset and Path Resolution
 - [ ] Update `resolve_asset_path()` in `utils.py` for dual-mode (PyInstaller + pip)
