+++ src/whisper_key/assets/sounds/record_stop.wav
