+++ src/whisper_key/assets/tray_idle.png
