+++ src/whisper_key/assets/tray_processing.png
