+++ src/whisper_key/assets/tray_recording.png
