+++ src/whisper_key/clipboard_manager.py
