+++ src/whisper_key/instance_manager.py
