+++ src/whisper_key/state_manager.py
